refactor: Explode ViewStack for individual components to be imported
