+++ react/ViewStack/context.js
@@ -1,0 +1,4 @@
+import React, { useContext } from 'react'
+
+export const ViewStackContext = React.createContext()
+export const useViewStack = () => useContext(ViewStackContext)
